refactor: Http Request parsing을 Request로 이동
